+++ src/senaite/core/catalog/utils.py
